[refactor/#3] 패키지 구조 리팩토링
- 기존 components -> common으로 명칭
- Activity, Screen 관련 파일들을 별도의 presentation 폴더로 이동
- BottomNavigation 화면 별로 패키지 구분
- util 폴더를 상위로 이동
