+++ app/src/main/AndroidManifest.xml
@@ -13,23 +13,23 @@
         android:theme="@style/Theme.ATSOPTANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".ui.screens.main.MyActivity"
+            android:name=".presentation.home.MyActivity"
             android:exported="false"
             android:label="@string/title_activity_my"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".ui.screens.signup.SignUpActivity"
+            android:name=".presentation.auth.signup.SignUpActivity"
             android:exported="false"
             android:label="@string/title_activity_signup"
             android:theme="@style/Theme.ATSOPTANDROID"
             android:windowSoftInputMode="adjustResize"/>
         <activity
-            android:name=".ui.screens.login.LoginActivity"
+            android:name=".presentation.auth.login.LoginActivity"
             android:exported="true"
             android:label="@string/title_activity_login"
             android:theme="@style/Theme.ATSOPTANDROID"/>
         <activity
-            android:name=".ui.screens.main.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.ATSOPTANDROID">
             <intent-filter>
