--- app/src/main/java/org/sopt/at/icon/BackIcon.kt
@@ -1,50 +1,0 @@
-/*
-* Converted using https://composables.com/svgtocompose
-*/
-
-import androidx.compose.foundation.Image
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.PathFillType
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.StrokeJoin
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.graphics.vector.path
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.unit.dp
-
-public val BackIcon: ImageVector
-	get() {
-		if (_back != null) {
-			return _back!!
-		}
-		_back = ImageVector.Builder(
-            name = "Ic_back",
-            defaultWidth = 24.dp,
-            defaultHeight = 24.dp,
-            viewportWidth = 24f,
-            viewportHeight = 24f
-        ).apply {
-			path(
-    			fill = null,
-    			fillAlpha = 1.0f,
-    			stroke = SolidColor(Color(0xFFFFFFFF)),
-    			strokeAlpha = 1.0f,
-    			strokeLineWidth = 2f,
-    			strokeLineCap = StrokeCap.Round,
-    			strokeLineJoin = StrokeJoin.Round,
-    			strokeLineMiter = 1.0f,
-    			pathFillType = PathFillType.NonZero
-			) {
-				moveTo(11f, 21f)
-				lineTo(2f, 12f)
-				lineTo(11f, 3.5f)
-			}
-		}.build()
-		return _back!!
-	}
-
-private var _back: ImageVector? = null
