+++ app/src/main/java/org/sopt/at/presentation/main/MainActivity.kt
@@ -1,0 +1,21 @@
+package org.sopt.at.presentation.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
+
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+
+        setContent {
+            ATSOPTANDROIDTheme {
+                MainScreen()
+            }
+        }
+    }
+}
