+++ app/src/main/java/org/sopt/at/presentation/shorts/ShortsScreen.kt
@@ -1,0 +1,34 @@
+package org.sopt.at.presentation.shorts
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import kotlinx.serialization.Serializable
+import org.sopt.at.presentation.main.navigation.BottomNavItem
+
+@Serializable
+data object Shorts
+
+@Composable
+fun ShortsScreen(paddingValues: PaddingValues,
+) {
+    Column(
+        modifier = Modifier
+            .padding(paddingValues)
+            .fillMaxSize(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        Text(
+            text = BottomNavItem.SHORTS.label,
+            style = MaterialTheme.typography.titleMedium
+        )
+    }
+}
