--- app/src/main/java/org/sopt/at/ui/components/appbar/CommonAppBar.kt
@@ -1,34 +1,0 @@
-package org.sopt.at.ui.components.appbar
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.vectorResource
-import org.sopt.at.R
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CommonTopAppBar(
-    modifier: Modifier = Modifier,
-    onBackClick: () -> Unit,
-    trailingIcon: @Composable () -> Unit
-) {
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        IconButton(onClick = onBackClick) {
-            Icon(
-                imageVector = ImageVector.vectorResource(id = R.drawable.ic_back),
-                contentDescription = null
-            )
-        }
-        trailingIcon
-    }
-}
