--- app/src/main/java/org/sopt/at/ui/screens/home/component/HomeTopBar.kt
@@ -1,57 +1,0 @@
-package org.sopt.at.ui.screens.home.component
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.unit.dp
-import org.sopt.at.R
-
-@Composable
-fun HomeTopBar(modifier: Modifier = Modifier) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(
-                vertical = 10.dp,
-                horizontal = dimensionResource(R.dimen.screen_padding_horizontal)
-            ),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Image(
-            painter = painterResource(R.drawable.ic_tving_text_logo),
-            contentDescription = null,
-            modifier = Modifier.width(80.dp)
-        )
-        Spacer(Modifier.weight(1f))
-        Image(
-            imageVector = ImageVector.vectorResource(R.drawable.ic_chromcast),
-            contentDescription = null,
-            modifier = Modifier.size(24.dp)
-        )
-        Spacer(Modifier.width(18.dp))
-        Image(
-            painter = painterResource(R.drawable.img_profile),
-            contentDescription = null,
-            modifier = Modifier
-                .size(24.dp)
-                .clickable(
-                    enabled = true,
-                    onClick = {
-                        //TODO: My 화면으로 이동
-                    }
-                )
-        )
-    }
-}
