--- app/src/main/java/org/sopt/at/ui/screens/main/LiveScreen.kt
@@ -1,34 +1,0 @@
-package org.sopt.at.ui.screens.main
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import kotlinx.serialization.Serializable
-import org.sopt.at.ui.screens.main.navigation.BottomNavItem
-
-@Serializable
-data object Live
-
-@Composable
-fun LiveScreen(paddingValues: PaddingValues,
-) {
-    Column(
-        modifier = Modifier
-            .padding(paddingValues)
-            .fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Text(
-            text = BottomNavItem.LIVE.label,
-            style = MaterialTheme.typography.titleMedium
-        )
-    }
-}
