--- app/src/main/java/org/sopt/at/ui/util/RippleControl.kt
@@ -1,14 +1,0 @@
-package org.sopt.at.ui.util
-
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.LocalRippleConfiguration
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun DisableRippleEffect(content: @Composable () -> Unit) {
-    CompositionLocalProvider(LocalRippleConfiguration provides null) {
-        content()
-    }
-}
