+++ app/src/main/java/org/sopt/at/util/RippleControl.kt
@@ -1,0 +1,14 @@
+package org.sopt.at.util
+
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.LocalRippleConfiguration
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun DisableRippleEffect(content: @Composable () -> Unit) {
+    CompositionLocalProvider(LocalRippleConfiguration provides null) {
+        content()
+    }
+}
